[SIG-1006] [SIG-1006] moved kto form in own component
