+++ src/signals/incident/containers/KtoContainer/components/KtoForm/index.test.js
@@ -1,0 +1,10 @@
+// import React from 'react';
+// import { shallow } from 'enzyme';
+
+// import KtoForm from 'index';
+
+describe('<KtoForm />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(true);
+  });
+});
