+++ src/signals/incident/containers/KtoContainer/style.scss
@@ -1,11 +1,4 @@
 @import "config.scss";
 
 .kto-container {
-  .row + .row {
-    margin-top: 20px !important;
-  }
-
-  .row {
-    margin-top: 20px !important;
-  }
 }
